requirements.txt
Django~=3.1.7
django-leaflet~=0.27.1
geopy~=2.1.0
folium~=0.12.1